requirements.txt
openai
dotenv
requests
flask
pyjwt
supabase
flask_cors
pymupdf